requirements.txt
appdirs==1.4.4
asgiref==3.4.1
autopep8==1.5.5
backports.entry-points-selectable==1.1.0
certifi==2021.10.8
cffi==1.14.6
coverage==5.5
cryptography==3.4.8
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2.6
django-ckeditor==6.1.0
django-cors-headers==3.8.0
django-crispy-forms==1.13.0
django-filter==21.1
django-js-asset==1.2.2
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-webpack-loader==1.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
Faker==6.5.2
filelock==3.3.2
fontawesomefree==6.0.0
gunicorn==20.0.4
mysqlclient==2.1.0
Pillow==8.4.0
pipenv==2020.11.15
platformdirs==2.4.0
pycodestyle==2.8.0
pycparser==2.20
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
validate-email==1.3
virtualenv==20.10.0
virtualenv-clone==0.5.7
